services/verification/requirements.txt
anthropic==0.39.0
beautifulsoup4==4.12.3
fastapi==0.110.3
httpx==0.27.2
pydantic==2.9.2
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.0.8
uvicorn==0.30.6
